Include intermediate directories when tarring
