+++ cargo/file_bundler.go
@@ -3,11 +3,59 @@ package cargo
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"time"
 )
+
+type File struct {
+	io.ReadCloser
+
+	Name string
+	Info os.FileInfo
+}
+
+type FileInfo struct {
+	name  string
+	size  int
+	mode  os.FileMode
+	mtime time.Time
+}
+
+func NewFileInfo(name string, size int, mode os.FileMode, mtime time.Time) FileInfo {
+	return FileInfo{
+		name:  name,
+		size:  size,
+		mode:  mode,
+		mtime: mtime,
+	}
+}
+
+func (fi FileInfo) Name() string {
+	return fi.name
+}
+
+func (fi FileInfo) Size() int64 {
+	return int64(fi.size)
+}
+
+func (fi FileInfo) Mode() os.FileMode {
+	return fi.mode
+}
+
+func (fi FileInfo) ModTime() time.Time {
+	return fi.mtime
+}
+
+func (fi FileInfo) IsDir() bool {
+	return fi.Mode().IsDir()
+}
+
+func (fi FileInfo) Sys() interface{} {
+	return nil
+}
 
 type FileBundler struct{}
 
@@ -17,6 +65,7 @@ func NewFileBundler() FileBundler {
 
 func (b FileBundler) Bundle(root string, paths []string, config Config) ([]File, error) {
 	var files []File
+
 	for _, path := range paths {
 		file := File{Name: path}
 
@@ -44,7 +93,6 @@ func (b FileBundler) Bundle(root string, paths []string, config Config) ([]File,
 
 			file.ReadCloser = fd
 			file.Info = info
-
 		}
 
 		files = append(files, file)
